Refactor/pydantic category webhook
